+++ src/main/java/com/nexmo/client/incoming/CallDirection.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 Nexmo Inc
+ * Copyright (c) 2011-2018 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,34 +19,27 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.voice;
+package com.nexmo.client.incoming;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public enum MachineDetection {
-    CONTINUE, HANGUP, UNKNOWN;
+public enum CallDirection {
+    OUTBOUND, INBOUND, UNKNOWN;
 
-    private static final Map<String, MachineDetection> MACHINE_DETECTION_INDEX = new HashMap<>();
+    private static final Map<String, CallDirection> CALL_DIRECTION_INDEX = new HashMap<>();
 
     static {
-        for (MachineDetection machineDetection : MachineDetection.values()) {
-            MACHINE_DETECTION_INDEX.put(machineDetection.name(), machineDetection);
+        for (CallDirection callDirection : CallDirection.values()) {
+            CALL_DIRECTION_INDEX.put(callDirection.name(), callDirection);
         }
     }
 
-    @JsonValue
-    @Override
-    public String toString() {
-        return name().toLowerCase();
-    }
-
     @JsonCreator
-    public static MachineDetection fromString(String name) {
-        MachineDetection foundMachineDetection = MACHINE_DETECTION_INDEX.get(name.toUpperCase());
-        return (foundMachineDetection != null) ? foundMachineDetection : UNKNOWN;
+    public static CallDirection fromString(String name) {
+        CallDirection foundCallDirection = CALL_DIRECTION_INDEX.get(name.toUpperCase());
+        return (foundCallDirection != null) ? foundCallDirection : UNKNOWN;
     }
 }