+++ src/test/java/com/nexmo/client/incoming/InputEventTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 Nexmo Inc
+ * Copyright (c) 2011-2018 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,23 +19,33 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.voice.ncco;
+package com.nexmo.client.incoming;
 
 import org.junit.Test;
 
+import java.util.Calendar;
+import java.util.GregorianCalendar;
+import java.util.TimeZone;
+
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
-public class RecordingFormatTest {
+public class InputEventTest {
     @Test
-    public void testFromString() throws Exception {
-        assertEquals(RecordingFormat.MP3, RecordingFormat.fromString("mp3"));
-        assertEquals(RecordingFormat.WAV, RecordingFormat.fromString("wav"));
-        assertEquals(RecordingFormat.UNKNOWN, RecordingFormat.fromString("test unknown"));
-    }
+    public void testDeserializeInputEvent() {
+        String json = "{\n" + "  \"uuid\": \"aaaaaaaa-bbbb-cccc-dddd-0123456789ab\",\n"
+                + "  \"conversation_uuid\": \"bbbbbbbb-cccc-dddd-eeee-0123456789ab\",\n" + "  \"timed_out\": true,\n"
+                + "  \"dtmf\": \"1234\",\n" + "  \"timestamp\": \"2020-01-01T14:00:00.000Z\"\n" + "}";
 
-    @Test
-    public void testToString() throws Exception {
-        assertEquals("mp3", RecordingFormat.MP3.toString());
-        assertEquals("wav", RecordingFormat.WAV.toString());
+        InputEvent inputEvent = InputEvent.fromJson(json);
+        assertEquals("aaaaaaaa-bbbb-cccc-dddd-0123456789ab", inputEvent.getUuid());
+        assertEquals("bbbbbbbb-cccc-dddd-eeee-0123456789ab", inputEvent.getConversationUuid());
+        assertTrue(inputEvent.isTimedOut());
+        assertEquals("1234", inputEvent.getDtmf());
+
+        Calendar calendar = new GregorianCalendar(2020, Calendar.JANUARY, 1, 14, 0, 0);
+        calendar.setTimeZone(TimeZone.getTimeZone("UTC"));
+
+        assertEquals(calendar.getTime(), inputEvent.getTimestamp());
     }
 }